+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Ignition Launch 1.X.X
 
+1.  Use the new GUI API of ign-gazebo. This adds support for saving worlds to SDFormat from the GUI.
+    * [Pull Request 19](https://github.com/ignitionrobotics/ign-launch/pull/19)
+
 ### Ignition Launch 1.4.1 (2019-12-05)
 
 1. Unversioned lib name for cmds