+++ Changelog.md
@@ -3,7 +3,7 @@
 ### Ignition Launch 1.X.X
 
 1. Added ability to set the publication rate of the websocket server.
-    * [Pull Request 17](https://bitbucket.org/ignitionrobotics/ign-launch/pull-requests/17)
+    * [Pull Request 17](https://github.com/ignitionrobotics/ign-launch/pull/17)
 
 ### Ignition Launch 1.4.1 (2019-12-05)