+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### Ignition Launch 3.X.X
 
+
 ### Ignition Launch 3.0.0 (20XX-XX-XX)
 
 1. Depend on ign-gazebo4, ign-gui4